adding the variable length matrix, for cuda cpy

+++ Sequential/matrix_util.c
@@ -19,12 +19,12 @@ matrix_t* roll_matrix_list(matrix_list_t* list)
 		vector_size += list->matrix_list[i]->rows * list->matrix_list[i]->cols;
 	}
 	matrix_t* vector = matrix_constructor(1, vector_size);
-	double* current_index = vector->matrix;
+	float* current_index = vector->matrix;
 
 	for(i=0; i<list->num; i++)
 	{
 		unsigned int matrix_size = list->matrix_list[i]->rows * list->matrix_list[i]->cols;
-		memcpy(current_index, list->matrix_list[i]->matrix, matrix_size * sizeof(double));
+		memcpy(current_index, list->matrix_list[i]->matrix, matrix_size * sizeof(float));
 		current_index = current_index + matrix_size;
 	}
 	return vector;
@@ -36,7 +36,7 @@ matrix_list_t* unroll_matrix_list(matrix_t* vector, int num, unsigned int sizes[
 
 	matrix_list_t* list = matrix_list_constructor(num);
 
-	double* current_index = vector->matrix;
+	float* current_index = vector->matrix;
 	unsigned int i;
 
 	for(i=0; i<num; i++)
@@ -44,7 +44,7 @@ matrix_list_t* unroll_matrix_list(matrix_t* vector, int num, unsigned int sizes[
 		list->matrix_list[i] = matrix_constructor(sizes[i][ROW_INDEX], sizes[i][COL_INDEX]);
 
 		unsigned int matrix_size = sizes[i][ROW_INDEX] * sizes[i][COL_INDEX];
-		memcpy(list->matrix_list[i]->matrix, current_index, matrix_size * sizeof(double));
+		memcpy(list->matrix_list[i]->matrix, current_index, matrix_size * sizeof(float));
 		current_index = current_index + matrix_size;
 	}
